Add "Print Pull List" button and functionality
Implemented a new button "Print Pull List" in the motor views to allow users to print a pull list of available and listable products. Added related logic in the motor model to filter and generate pull list reports and extended the printnode interface to include the "pull list" job type.